Test files
These Python scripts are to be tested and implemented in their respective master files. The AITest files contain code, which will build a basic GUI with tkinter. It will have 2 squares - one controlled by the player and one by an AI. This code will be improved and used as a base for the characters in the VRS Pac-man game. Also contains an .ev3 bluetooth file, which makes one brick communicate with another.

The main difference between the two AI scripts is that all AI functions are only defined in the first one - the second square will not move since it's fucntions aren't called. They are called in the second version, so the AI is fully functional.

Credits:
AI Test files: Teodor Atanasov and Deyan Bozhilov
Bluetooth file: Nathan Boamah